+++ baybe/constraints/discrete.py
@@ -5,13 +5,16 @@ from functools import reduce
 from typing import Any, cast
 
 import pandas as pd
+import polars as pl
 from attr import define, field
 from attr.validators import in_, min_len
 
 from baybe.constraints.base import DiscreteConstraint
 from baybe.constraints.conditions import (
     Condition,
+    SubSelectionCondition,
     ThresholdCondition,
+    _threshold_operators,
     _valid_logic_combiners,
 )
 from baybe.serialization import (
@@ -42,6 +45,29 @@ class DiscreteExcludeConstraint(DiscreteConstraint):
         res = reduce(_valid_logic_combiners[self.combiner], satisfied)
         return data.index[res]
 
+    def to_polars(self) -> pl.Expr:
+        """Translate the conditions to polars expression for filtering.
+
+        Returns:
+            The Polars Expr object to pass as an argument to filter().
+        """
+        satisfied: list[pl.Expr] = []
+        for k, cond in enumerate(self.conditions):
+            # SubSelectionCondition requires parameter to generate an expression
+            if isinstance(cond, SubSelectionCondition):
+                satisfied.append(cond.to_polars(self.parameters[k]))
+            else:  # Assuming we have only two classes for condition
+                op = cond.to_polars(param=None)
+                satisfied.append(op(pl.col(self.parameters[k])))
+
+        # to prevent errors when we have empty condition list
+        expr = reduce(_valid_logic_combiners[self.combiner], satisfied)
+        # Negate the expression, because Polars' filter keeps the rows
+        # the expression satisfies
+        expr = expr.not_()
+
+        return expr
+
 
 @define
 class DiscreteSumConstraint(DiscreteConstraint):
@@ -60,6 +86,16 @@ class DiscreteSumConstraint(DiscreteConstraint):
 
         return data.index[mask_bad]
 
+    def to_polars(self) -> pl.Expr:
+        """Translate the conditions to polars expression for filtering.
+
+        Returns:
+            The Polars Expr object to pass as an argument to filter().
+
+        """
+        op = self.condition.to_polars()
+        return op(pl.sum_horizontal(self.parameters))
+
 
 @define
 class DiscreteProductConstraint(DiscreteConstraint):
@@ -77,6 +113,22 @@ class DiscreteProductConstraint(DiscreteConstraint):
         mask_bad = ~self.condition.evaluate(evaluate_data)
 
         return data.index[mask_bad]
+
+    def to_polars(self) -> pl.Expr:
+        """Translate the conditions to polars expression for filtering.
+
+        Returns:
+            The Polars Expr object to pass as an argument to filter().
+
+        """
+        op = _threshold_operators[self.condition.operator]
+        # Identify columns present in both self.parameters and df.columns
+
+        # Get the product of columns using reduce function and with an alias "prod"
+        expr = pl.reduce(lambda acc, x: acc * x, pl.col(self.parameters)).alias("prod")
+
+        # Apply the threshold operator on expr and the condition threshold
+        return op(expr, self.condition.threshold)
 
 
 class DiscreteNoLabelDuplicatesConstraint(DiscreteConstraint):